requirements.txt
anyio==3.5.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
dnspython==2.2.0
fastapi==0.72.0
h11==0.12.0
hvac==0.11.2
idna==3.3
ifaddr==0.1.7
langdetect==1.0.9
numpy==1.22.1
pandas==1.3.5
Pillow==9.0.0
pydantic==1.9.0
pytesseract==0.3.7
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
urllib3==1.26.9
uvicorn==0.17.5
